Syntactic Sugar for Filter Construction

These functions complements mlr_filters with a function in the spirit of mlr3::mlr_sugar.

flt(.key, ...)

flts(.keys, ...)

Arguments

.key
(character(1))
Key passed to the respective dictionary to retrieve the object.

...
(named list())
Named arguments passed to the constructor, to be set as parameters in the paradox::ParamSet, or to be set as public field. See mlr3misc::dictionary_sugar_get() for more details.

.keys
(character())
Keys passed to the respective dictionary to retrieve multiple objects.

Value

Filter.

Examples

flt("correlation", method = "kendall")

#> <FilterCorrelation:correlation>
#> Task Types: regr
#> Task Properties: -
#> Packages: stats
#> Feature types: integer, numeric

flts(c("mrmr", "jmim"))

#> [[1]]
#> <FilterMRMR:mrmr>
#> Task Types: classif, regr
#> Task Properties: -
#> Packages: praznik
#> Feature types: integer, numeric, factor, ordered
#> 
#> [[2]]
#> <FilterJMIM:jmim>
#> Task Types: classif, regr
#> Task Properties: -
#> Packages: praznik
#> Feature types: integer, numeric, factor, ordered
#>