Dictionary of Filters

A simple Dictionary storing objects of class Filter. Each Filter has an associated help page, see mlr_filters_[id].

This dictionary can get populated with additional filters by add-on packages.

For a more convenient way to retrieve and construct filters, see flt().

mlr_filters

Format

R6Class object

Usage

See Dictionary.

See also

Examples

mlr_filters$keys()

#>  [1] "anova"            "auc"              "carscore"         "cmim"            
#>  [5] "correlation"      "disr"             "importance"       "information_gain"
#>  [9] "jmi"              "jmim"             "kruskal_test"     "mim"             
#> [13] "mrmr"             "njmim"            "performance"      "variance"        

as.data.table(mlr_filters)

#>                  key    task_type task_properties  param_set
#>  1:            anova      classif                 <ParamSet>
#>  2:              auc      classif        twoclass <ParamSet>
#>  3:         carscore         regr                 <ParamSet>
#>  4:             cmim classif,regr                 <ParamSet>
#>  5:      correlation         regr                 <ParamSet>
#>  6:             disr      classif                 <ParamSet>
#>  7:       importance      classif                 <ParamSet>
#>  8: information_gain classif,regr                 <ParamSet>
#>  9:              jmi      classif                 <ParamSet>
#> 10:             jmim      classif                 <ParamSet>
#> 11:     kruskal_test      classif                 <ParamSet>
#> 12:              mim      classif                 <ParamSet>
#> 13:             mrmr classif,regr                 <ParamSet>
#> 14:            njmim      classif                 <ParamSet>
#> 15:      performance      classif                 <ParamSet>
#> 16:         variance classif,regr                 <ParamSet>
#>                                feature_types      packages
#>  1:                          integer,numeric         stats
#>  2:                          integer,numeric  mlr3measures
#>  3:                                  numeric          care
#>  4:           integer,numeric,factor,ordered       praznik
#>  5:                          integer,numeric         stats
#>  6:           integer,numeric,factor,ordered       praznik
#>  7:   logical,integer,numeric,factor,ordered         rpart
#>  8:           integer,numeric,factor,ordered FSelectorRcpp
#>  9:           integer,numeric,factor,ordered       praznik
#> 10:           integer,numeric,factor,ordered       praznik
#> 11:                          integer,numeric         stats
#> 12:           integer,numeric,factor,ordered       praznik
#> 13: numeric,factor,integer,character,logical       praznik
#> 14:           integer,numeric,factor,ordered       praznik
#> 15:   logical,integer,numeric,factor,ordered         rpart
#> 16:                          integer,numeric         stats

mlr_filters$get("mim")

#> <FilterMIM:mim>
#> Task Types: classif
#> Task Properties: -
#> Packages: praznik
#> Feature types: integer, numeric, factor, ordered

flt("anova")

#> <FilterAnova:anova>
#> Task Types: classif
#> Task Properties: -
#> Packages: stats
#> Feature types: integer, numeric