Burota Filter

Filter using the Boruta algorithm for feature selection. If keep = "tentative", confirmed and tentative features are returned. Note that there is no ordering in the selected features. Selected features get a score of 1, deselected features get a score of 0. The order of selected features is random. In combination with mlr3pipelines, only the filter criterion cutoff makes sense.

References

Kursa MB, Rudnicki WR (2010). “Feature Selection with the Boruta Package.” Journal of Statistical Software, 36(11), 1-13.

See also

Super class

mlr3filters::Filter -> FilterBoruta

Methods

Public methods

FilterBoruta$new()

FilterBoruta$clone()

Inherited methods

mlr3filters::Filter$calculate()

mlr3filters::Filter$format()

mlr3filters::Filter$help()

mlr3filters::Filter$print()

Method new()

Creates a new instance of this R6 class.

Usage

FilterBoruta$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterBoruta$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \donttest{
  if (requireNamespace("Boruta")) {
   task = mlr3::tsk("sonar")
   filter = flt("boruta")
   filter$calculate(task)
   as.data.table(filter)
  }
#> Loading required namespace: Boruta
#>     feature score
#>      <char> <num>
#>  1:     V16     1
#>  2:     V15     1
#>  3:     V27     1
#>  4:      V4     1
#>  5:     V44     1
#>  6:      V9     1
#>  7:     V52     1
#>  8:     V17     1
#>  9:     V31     1
#> 10:     V26     1
#> 11:     V13     1
#> 12:     V49     1
#> 13:     V46     1
#> 14:     V19     1
#> 15:     V18     1
#> 16:     V10     1
#> 17:     V23     1
#> 18:     V45     1
#> 19:     V36     1
#> 20:      V1     1
#> 21:     V47     1
#> 22:     V22     1
#> 23:      V5     1
#> 24:     V51     1
#> 25:     V20     1
#> 26:     V37     1
#> 27:     V48     1
#> 28:     V35     1
#> 29:     V21     1
#> 30:     V28     1
#> 31:     V12     1
#> 32:     V11     1
#> 33:      V7     0
#> 34:     V29     0
#> 35:     V55     0
#> 36:     V42     0
#> 37:     V40     0
#> 38:     V34     0
#> 39:     V32     0
#> 40:     V60     0
#> 41:     V25     0
#> 42:     V39     0
#> 43:      V3     0
#> 44:     V56     0
#> 45:     V24     0
#> 46:     V54     0
#> 47:     V14     0
#> 48:     V53     0
#> 49:     V30     0
#> 50:      V2     0
#> 51:     V41     0
#> 52:     V57     0
#> 53:      V6     0
#> 54:      V8     0
#> 55:     V43     0
#> 56:     V33     0
#> 57:     V58     0
#> 58:     V38     0
#> 59:     V50     0
#> 60:     V59     0
#>     feature score
# }