Double Input Symmetrical Relevance Filter

Double input symmetrical relevance filter calling praznik::DISR() from package praznik.

This filter supports partial scoring (see Filter).

Usage,NULL

Details

As the scores calculated by the praznik package are not monotone due to the greedy forward fashion, the returned scores simply reflect the selection order: 1, (k-1)/k, ..., 1/k where k is the number of selected features.

Threading is disabled by default (hyperparameter threads is set to 1). Set to a number >= 2 to enable threading, or to 0 for auto-detecting the number of available cores.

References

Kursa MB (2021). “Praznik: High performance information-based feature selection.” SoftwareX, 16, 100819. doi: 10.1016/j.softx.2021.100819 .

For a benchmark of filter methods:

Bommert A, Sun X, Bischl B, Rahnenführer J, Lang M (2020). “Benchmark for filter methods for feature selection in high-dimensional classification data.” Computational Statistics & Data Analysis, 143, 106839. doi: 10.1016/j.csda.2019.106839 .

See also

Super class

mlr3filters::Filter -> FilterDISR

Methods

Public methods

FilterDISR$new()

FilterDISR$clone()

Inherited methods

mlr3filters::Filter$calculate()

mlr3filters::Filter$format()

mlr3filters::Filter$help()

mlr3filters::Filter$print()

Method new()

Create a FilterDISR object.

Usage

FilterDISR$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterDISR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = mlr3::tsk("iris")
filter = flt("disr")
filter$calculate(task)
as.data.table(filter)
#>         feature     score
#> 1:  Petal.Width 1.0000000
#> 2: Petal.Length 0.6666667
#> 3: Sepal.Length 0.3333333
#> 4:  Sepal.Width 0.0000000