Mutual Information Maximization Filter

Conditional mutual information based feature selection filter calling praznik::MIM() in package praznik.

This filter supports partial scoring (see Filter).

Details

As the scores calculated by the praznik package are not monotone due to the greedy forward fashion, the returned scores simply reflect the selection order: 1, (k-1)/k, ..., 1/k where k is the number of selected features.

Threading is disabled by default (hyperparameter threads is set to 1). Set to a number >= 2 to enable threading, or to 0 for auto-detecting the number of available cores.

References

Kursa MB (2021). “Praznik: High performance information-based feature selection.” SoftwareX, 16, 100819. doi:10.1016/j.softx.2021.100819 .

For a benchmark of filter methods:

Bommert A, Sun X, Bischl B, Rahnenführer J, Lang M (2020). “Benchmark for filter methods for feature selection in high-dimensional classification data.” Computational Statistics & Data Analysis, 143, 106839. doi:10.1016/j.csda.2019.106839 .

See also

Super class

mlr3filters::Filter -> FilterMIM

Methods

Public methods

FilterMIM$new()

FilterMIM$clone()

Inherited methods

mlr3filters::Filter$calculate()

mlr3filters::Filter$format()

mlr3filters::Filter$help()

mlr3filters::Filter$print()

Method new()

Create a FilterMIM object.

Usage

FilterMIM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterMIM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("praznik")) {
  task = mlr3::tsk("iris")
  filter = flt("mim")
  filter$calculate(task, nfeat = 2)
  as.data.table(filter)
}
#>         feature score
#>          <char> <num>
#> 1:  Petal.Width     1
#> 2: Petal.Length     0
#> 3: Sepal.Length    NA
#> 4:  Sepal.Width    NA

if (mlr3misc::require_namespaces(c("mlr3pipelines", "rpart", "praznik"), quietly = TRUE)) {
  library("mlr3pipelines")
  task = mlr3::tsk("spam")

  # Note: `filter.frac` is selected randomly and should be tuned.

  graph = po("filter", filter = flt("mim"), filter.frac = 0.5) %>>%
    po("learner", mlr3::lrn("classif.rpart"))

  graph$train(task)
}
#> $classif.rpart.output
#> NULL
#>