Dictionary of Filters

A simple mlr3misc::Dictionary storing objects of class Filter. Each Filter has an associated help page, see mlr_filters_[id].

This dictionary can get populated with additional filters by add-on packages.

For a more convenient way to retrieve and construct filters, see flt().

Usage

mlr_filters

Format

R6::R6Class object

Usage

See mlr3misc::Dictionary.

See also

Examples

mlr_filters$keys()
#>  [1] "anova"             "auc"               "boruta"           
#>  [4] "carscore"          "carsurvscore"      "cmim"             
#>  [7] "correlation"       "disr"              "ensemble"         
#> [10] "find_correlation"  "importance"        "information_gain" 
#> [13] "jmi"               "jmim"              "kruskal_test"     
#> [16] "mim"               "mrmr"              "njmim"            
#> [19] "performance"       "permutation"       "relief"           
#> [22] "selected_features" "univariate_cox"    "variance"         
as.data.table(mlr_filters)
#> Error in assert_list(filters, types = "Filter", min.len = 1): argument "filters" is missing, with no default
mlr_filters$get("mim")
#> 
#> ── <FilterMIM> mim: Mutual Information Maximization ────────────────────────────
#> • Task Types: classif and regr
#> • Properties: -
#> • Task Properties:
#> • Packages: praznik
#> • Feature types: integer, numeric, factor, and ordered
flt("anova")
#> 
#> ── <FilterAnova> anova: ANOVA F-Test ───────────────────────────────────────────
#> • Task Types: classif
#> • Properties: -
#> • Task Properties:
#> • Packages: stats
#> • Feature types: integer and numeric