Dictionary of Filters

A simple Dictionary storing objects of class Filter. Each Filter has an associated help page, see mlr_filters_[id].

This dictionary can get populated with additional filters by add-on packages.

For a more convenient way to retrieve and construct filters, see flt().

mlr_filters

Format

R6Class object

Usage

See Dictionary.

See also

Examples

mlr_filters$keys()

#>  [1] "anova"            "auc"              "carscore"         "cmim"            
#>  [5] "correlation"      "disr"             "find_correlation" "importance"      
#>  [9] "information_gain" "jmi"              "jmim"             "kruskal_test"    
#> [13] "mim"              "mrmr"             "njmim"            "performance"     
#> [17] "permutation"      "relief"           "variance"        

as.data.table(mlr_filters)

#>                  key    task_type task_properties      param_set
#>  1:            anova      classif                 <ParamSet[40]>
#>  2:              auc      classif        twoclass <ParamSet[40]>
#>  3:         carscore         regr                 <ParamSet[40]>
#>  4:             cmim classif,regr                 <ParamSet[40]>
#>  5:      correlation         regr                 <ParamSet[40]>
#>  6:             disr      classif                 <ParamSet[40]>
#>  7: find_correlation classif,regr                 <ParamSet[40]>
#>  8:       importance      classif                 <ParamSet[40]>
#>  9: information_gain classif,regr                 <ParamSet[40]>
#> 10:              jmi      classif                 <ParamSet[40]>
#> 11:             jmim      classif                 <ParamSet[40]>
#> 12:     kruskal_test      classif                 <ParamSet[40]>
#> 13:              mim      classif                 <ParamSet[40]>
#> 14:             mrmr      classif                 <ParamSet[40]>
#> 15:            njmim      classif                 <ParamSet[40]>
#> 16:      performance      classif                 <ParamSet[40]>
#> 17:      permutation      classif                 <ParamSet[40]>
#> 18:           relief classif,regr                 <ParamSet[40]>
#> 19:         variance classif,regr                 <ParamSet[40]>
#>                              feature_types           packages
#>  1:                        integer,numeric              stats
#>  2:                        integer,numeric       mlr3measures
#>  3:                                numeric               care
#>  4:         integer,numeric,factor,ordered            praznik
#>  5:                        integer,numeric              stats
#>  6:         integer,numeric,factor,ordered            praznik
#>  7:                        integer,numeric              stats
#>  8: logical,integer,numeric,factor,ordered              rpart
#>  9:         integer,numeric,factor,ordered      FSelectorRcpp
#> 10:         integer,numeric,factor,ordered            praznik
#> 11:         integer,numeric,factor,ordered            praznik
#> 12:                        integer,numeric              stats
#> 13:         integer,numeric,factor,ordered            praznik
#> 14:         integer,numeric,factor,ordered            praznik
#> 15:         integer,numeric,factor,ordered            praznik
#> 16: logical,integer,numeric,factor,ordered rpart,mlr3measures
#> 17: logical,integer,numeric,factor,ordered rpart,mlr3measures
#> 18:         integer,numeric,factor,ordered      FSelectorRcpp
#> 19:                        integer,numeric              stats

mlr_filters$get("mim")

#> <FilterMIM:mim>
#> Task Types: classif
#> Task Properties: -
#> Packages: praznik
#> Feature types: integer, numeric, factor, ordered

flt("anova")

#> <FilterAnova:anova>
#> Task Types: classif
#> Task Properties: -
#> Packages: stats
#> Feature types: integer, numeric