Dictionary of Filters

A simple Dictionary storing objects of class Filter. Each Filter has an associated help page, see mlr_filters_[id].

This dictionary can get populated with additional filters by add-on packages.

For a more convenient way to retrieve and construct filters, see flt().

Usage

mlr_filters

Format

R6Class object

Usage

See Dictionary.

See also

Examples

mlr_filters$keys()
#>  [1] "anova"             "auc"               "carscore"         
#>  [4] "cmim"              "correlation"       "disr"             
#>  [7] "find_correlation"  "importance"        "information_gain" 
#> [10] "jmi"               "jmim"              "kruskal_test"     
#> [13] "mim"               "mrmr"              "njmim"            
#> [16] "performance"       "permutation"       "relief"           
#> [19] "selected_features" "variance"         
as.data.table(mlr_filters)
#>                   key    task_type task_properties      param_set
#>  1:             anova      classif                 <ParamSet[42]>
#>  2:               auc      classif        twoclass <ParamSet[42]>
#>  3:          carscore         regr                 <ParamSet[42]>
#>  4:              cmim classif,regr                 <ParamSet[42]>
#>  5:       correlation         regr                 <ParamSet[42]>
#>  6:              disr classif,regr                 <ParamSet[42]>
#>  7:  find_correlation classif,regr                 <ParamSet[42]>
#>  8:        importance      classif                 <ParamSet[42]>
#>  9:  information_gain classif,regr                 <ParamSet[42]>
#> 10:               jmi classif,regr                 <ParamSet[42]>
#> 11:              jmim classif,regr                 <ParamSet[42]>
#> 12:      kruskal_test      classif                 <ParamSet[42]>
#> 13:               mim classif,regr                 <ParamSet[42]>
#> 14:              mrmr classif,regr                 <ParamSet[42]>
#> 15:             njmim classif,regr                 <ParamSet[42]>
#> 16:       performance      classif                 <ParamSet[42]>
#> 17:       permutation      classif                 <ParamSet[42]>
#> 18:            relief classif,regr                 <ParamSet[42]>
#> 19: selected_features      classif                 <ParamSet[42]>
#> 20:          variance           NA                 <ParamSet[42]>
#>                              feature_types                            packages
#>  1:                        integer,numeric                   mlr3filters,stats
#>  2:                        integer,numeric            mlr3filters,mlr3measures
#>  3:                                numeric                    mlr3filters,care
#>  4:         integer,numeric,factor,ordered                 mlr3filters,praznik
#>  5:                        integer,numeric                   mlr3filters,stats
#>  6:         integer,numeric,factor,ordered                 mlr3filters,praznik
#>  7:                        integer,numeric                   mlr3filters,stats
#>  8: logical,integer,numeric,factor,ordered              mlr3filters,mlr3,rpart
#>  9:         integer,numeric,factor,ordered           mlr3filters,FSelectorRcpp
#> 10:         integer,numeric,factor,ordered                 mlr3filters,praznik
#> 11:         integer,numeric,factor,ordered                 mlr3filters,praznik
#> 12:                        integer,numeric                   mlr3filters,stats
#> 13:         integer,numeric,factor,ordered                 mlr3filters,praznik
#> 14:         integer,numeric,factor,ordered                 mlr3filters,praznik
#> 15:         integer,numeric,factor,ordered                 mlr3filters,praznik
#> 16: logical,integer,numeric,factor,ordered mlr3filters,mlr3,rpart,mlr3measures
#> 17: logical,integer,numeric,factor,ordered mlr3filters,mlr3,rpart,mlr3measures
#> 18:         integer,numeric,factor,ordered           mlr3filters,FSelectorRcpp
#> 19: logical,integer,numeric,factor,ordered              mlr3filters,mlr3,rpart
#> 20:                        integer,numeric                   mlr3filters,stats
mlr_filters$get("mim")
#> <FilterMIM:mim>
#> Task Types: classif, regr
#> Task Properties: -
#> Packages: mlr3filters, praznik
#> Feature types: integer, numeric, factor, ordered
flt("anova")
#> <FilterAnova:anova>
#> Task Types: classif
#> Task Properties: -
#> Packages: mlr3filters, stats
#> Feature types: integer, numeric