Dictionary of Filters

A simple Dictionary storing objects of class Filter. Each Filter has an associated help page, see mlr_filters_[id].

This dictionary can get populated with additional filters by add-on packages.

For a more convenient way to retrieve and construct filters, see flt().

Usage

mlr_filters

Format

R6Class object

Usage

See Dictionary.

See also

Examples

mlr_filters$keys()
#>  [1] "anova"             "auc"               "carscore"         
#>  [4] "carsurvscore"      "cmim"              "correlation"      
#>  [7] "disr"              "find_correlation"  "importance"       
#> [10] "information_gain"  "jmi"               "jmim"             
#> [13] "kruskal_test"      "mim"               "mrmr"             
#> [16] "njmim"             "performance"       "permutation"      
#> [19] "relief"            "selected_features" "variance"         
as.data.table(mlr_filters)
#>                   key                                                    label
#>  1:             anova                                             ANOVA F-Test
#>  2:               auc                           Area Under the ROC Curve Score
#>  3:          carscore                   Correlation-Adjusted coRrelation Score
#>  4:      carsurvscore          Correlation-Adjusted coRrelation Survival Score
#>  5:              cmim      Minimal Conditional Mutual Information Maximization
#>  6:       correlation                                              Correlation
#>  7:              disr                       Double Input Symmetrical Relevance
#>  8:  find_correlation                                  Correlation-based Score
#>  9:        importance                                         Importance Score
#> 10:  information_gain                                         Information Gain
#> 11:               jmi                                 Joint Mutual Information
#> 12:              jmim            Minimal Joint Mutual Information Maximization
#> 13:      kruskal_test                                      Kruskal-Wallis Test
#> 14:               mim                          Mutual Information Maximization
#> 15:              mrmr                     Minimum Redundancy Maximal Relevancy
#> 16:             njmim Minimal Normalised Joint Mutual Information Maximization
#> 17:       performance                                   Predictive Performance
#> 18:       permutation                                        Permutation Score
#> 19:            relief                                                   RELIEF
#> 20: selected_features                               Embedded Feature Selection
#> 21:          variance                                                 Variance
#>                   key                                                    label
#>       task_types task_properties                          params
#>  1:      classif                                                
#>  2:      classif        twoclass                                
#>  3:         regr                         lambda,diagonal,verbose
#>  4:         surv                              maxIPCweight,denom
#>  5: classif,regr                                         threads
#>  6:         regr                                      use,method
#>  7: classif,regr                                         threads
#>  8: classif,regr                                      use,method
#>  9:      classif                                          method
#> 10: classif,regr                 type,equal,discIntegers,threads
#> 11: classif,regr                                         threads
#> 12: classif,regr                                         threads
#> 13:      classif                                       na.action
#> 14: classif,regr                                         threads
#> 15: classif,regr                                         threads
#> 16: classif,regr                                         threads
#> 17:      classif                                          method
#> 18:      classif                                 standardize,nmc
#> 19: classif,regr                      neighboursCount,sampleSize
#> 20:      classif                                          method
#> 21:           NA                                           na.rm
#>       task_types task_properties                          params
#>                                            feature_types          packages
#>  1:                                      integer,numeric             stats
#>  2:                                      integer,numeric      mlr3measures
#>  3:                                              numeric              care
#>  4:                                      integer,numeric carSurv,mlr3proba
#>  5:                       integer,numeric,factor,ordered           praznik
#>  6:                                      integer,numeric             stats
#>  7:                       integer,numeric,factor,ordered           praznik
#>  8:                                      integer,numeric             stats
#>  9: logical,integer,numeric,character,factor,ordered,...              mlr3
#> 10:                       integer,numeric,factor,ordered     FSelectorRcpp
#> 11:                       integer,numeric,factor,ordered           praznik
#> 12:                       integer,numeric,factor,ordered           praznik
#> 13:                                      integer,numeric             stats
#> 14:                       integer,numeric,factor,ordered           praznik
#> 15:                       integer,numeric,factor,ordered           praznik
#> 16:                       integer,numeric,factor,ordered           praznik
#> 17: logical,integer,numeric,character,factor,ordered,... mlr3,mlr3measures
#> 18: logical,integer,numeric,character,factor,ordered,... mlr3,mlr3measures
#> 19:                       integer,numeric,factor,ordered     FSelectorRcpp
#> 20: logical,integer,numeric,character,factor,ordered,...              mlr3
#> 21:                                      integer,numeric             stats
#>                                            feature_types          packages
mlr_filters$get("mim")
#> <FilterMIM:mim>: Mutual Information Maximization
#> Task Types: classif, regr
#> Task Properties: -
#> Packages: praznik
#> Feature types: integer, numeric, factor, ordered
flt("anova")
#> <FilterAnova:anova>: ANOVA F-Test
#> Task Types: classif
#> Task Properties: -
#> Packages: stats
#> Feature types: integer, numeric