Double Input Symmetrical Relevance Filter

Double input symmetrical relevance filter calling praznik::DISR() from package praznik.

This filter supports partial scoring (see Filter).

Details

As the scores calculated by the praznik package are not monotone due to the greedy forward fashion, the returned scores simply reflect the selection order: 1, (k-1)/k, ..., 1/k where k is the number of selected features.

Threading is disabled by default (hyperparameter threads is set to 1). Set to a number >= 2 to enable threading, or to 0 for auto-detecting the number of available cores.

See also

Super class

mlr3filters::Filter -> FilterDISR

Methods

Public methods

FilterDISR$new()

FilterDISR$clone()

Inherited methods

mlr3filters::Filter$calculate()

mlr3filters::Filter$format()

mlr3filters::Filter$help()

mlr3filters::Filter$print()

Method new()

Create a FilterDISR object.

Usage

FilterDISR$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

FilterDISR$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = mlr3::tsk("iris")
filter = flt("disr")
filter$calculate(task, nfeat = 2)
as.data.table(filter)

#>         feature score
#> 1:  Petal.Width   1.0
#> 2: Petal.Length   0.5
#> 3: Sepal.Length   0.0
#> 4:  Sepal.Width    NA